+++ src/Trarizon.Library/Helpers/UnsafeHelper.cs
@@ -7,10 +7,8 @@ public static class UnsafeHelper
     public static ref readonly TTo AsReadOnly<TFrom, TTo>(ref readonly TFrom source)
         => ref Unsafe.As<TFrom, TTo>(ref Unsafe.AsRef(in source));
 
-    public static unsafe nint Offset<T>(ref readonly T origin, ref readonly T target)
+    public static nint Offset<T>(ref readonly T origin, ref readonly T target)
     {
-#pragma warning disable CS8500
-        return Unsafe.ByteOffset(in origin, in target) / sizeof(T);
-#pragma warning restore CS8500
+        return Unsafe.ByteOffset(in origin, in target) / Unsafe.SizeOf<T>();
     }
 }